requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
asgiref==3.4.1; python_version >= '3.6'
dj-database-url==0.5.0
django-material-admin==1.6.37
django==3.2.9
djangorestframework==3.12.4
factory-boy==3.2.1
faker==9.8.3; python_version >= '3.6'
gunicorn==20.1.0
psycopg2==2.9.2
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==0.10.5
pytz==2021.3
setuptools==59.4.0; python_version >= '3.6'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlparse==0.4.2; python_version >= '3.5'
text-unidecode==1.3
whitenoise==5.3.0
